将构建逻辑迁移到 starter中
