+++ src/lib.rs
@@ -6,6 +6,7 @@ pub mod grpc;
 pub mod naming;
 pub mod raft;
 pub mod utils;
+pub mod starter;
 pub mod web_config;
 
 pub use inner_mem_cache::TimeoutSet;
